Added step to clone this repository when taking the CAPI route
Added step to clone this repository when taking the CAPI route. Also changed the `capi-package-f5-controller.sh` and `capi-package-f5-ipam-controller.sh` commands to include the path generated by cloning this directory. This will result in a clean directory structure and only the manifests to be applied will be in the `$CLUSTER_NAME` directory